tests/wpt/css-tests/selectors-3_dev/html/css3-selectors-lang-054.htm

<!DOCTYPE html>
<html lang="en"><head>
<meta charset="utf-8">
<title>[lang="cs-CZ"], lang="cs-Latn-CZ"</title>
<link href="mailto:ishida@w3.org" rel="author" title="Richard Ishida">
<link href="http://www.w3.org/TR/css3-selectors/#attribute-selectors" rel="help">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type="text/css">
.test div { width: 50px; }
#box[lang='cs-CZ'] { width: 100px; }
</style>
</head>
<body>



<div class="test"><div lang="cs-Latn-CZ" id="box">&nbsp;</div></div>


<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 50);
}, "A lang= value with language and region subtags will NOT match a lang attribute value with language, script and region subtags.");
</script>

<div id="log"></div>



</body></html>
